Changed name on filter.js to createTags.js and changed to functions
